refactor(adapter-server): extract batch_actions mutation into focused module (closes #237)
Pure refactor — no behavior change.

`build-schema.ts` was past the repo's 500-line guideline after PR #234
added the batch_actions subsystem inline. Move the dedicated GraphQL
types, the `serializeBatchResult` helper, and the `batch_actions`
mutation field into a new self-contained `build-batch-mutation.ts`
exposed as a `buildBatchMutationField({ commandLayer, transactionManager })`
factory.

- build-schema.ts: 1416 → 1186 lines (-230)
- build-batch-mutation.ts: 308 lines (new, under the 500 cap)
- No test files modified — existing batch tests cover the same behavior.
